Poll process for more reliable exit handling
We can't directly wait on the process launched by tmux since it isn't a
child process, so poll every 100ms to detect when it dies. This doesn't
seem to incur a noticeable performance penalty, but it is less than
ideal.

We may need to add a wrapper process like overmind's "launch" for a more
robust solution, this would help with ensuring we keep the tmux pane
open across restarts as well.