Pass through full request and response in proxy
Upstream server wasn't getting any headers, we were doing a fresh GET
request with just the URL from the request. Switch to passing the
request along to the upstream server with only the URL modified. Not
sure how safe it is to be mutating the received request like this, but
ownership should keep us on the straight and narrow to some degree.

Also pass the upstream response along verbatim, not sure why we were
splitting and recombining it.