Add rudimentary DNS server for MacOS
Add a barebones DNS server that always responds with localhost, based on
the implementation in puma-dev. This is meant for use with MacOS
/etc/resolvers/* domain specific server system, any requests to the
configured test TLD will be routed to localhost.

This turned out to be rather messy with the trust-dns library, it
requires a lot of boilerplate to get the basic functionality we need.