Switch Process from std Mutex to tokio RwLock
Using the blocking mutex from std is technically incorrect since all
operations in an async context are supposed to be non blocking. This
hasn't really been an issue so far (probably because the locks aren't
under much contention and blocking is minimal) but switch over just to
be safe.

Also switch to RwLock instead of a Mutex while we're at it, this should
speed things up slightly since we're pretty read-heavy.