Fix hang on opening pipe
We were using std's blocking file operations for deleting and opening
the pipe for some reason. Apparently named pipes block on open if there
is no reader, so we were hanging the entire runtime. Still not sure why
this is blocking in the first place since we should have a writer, but
switching to tokio's fs methods should ensure that we at least don't
block the runtime.